Move SourceTree outside PBXReference
