+++ subprojects/xcode-ide-kit/src/main/java/dev/nokee/xcode/objects/PBXProject.java
@@ -21,6 +21,7 @@ import dev.nokee.xcode.objects.configuration.XCConfigurationList;
 import dev.nokee.xcode.objects.files.PBXFileReference;
 import dev.nokee.xcode.objects.files.PBXGroup;
 import dev.nokee.xcode.objects.files.PBXReference;
+import dev.nokee.xcode.objects.files.PBXSourceTree;
 import dev.nokee.xcode.objects.targets.PBXTarget;
 
 import java.util.ArrayList;
@@ -40,7 +41,7 @@ public final class PBXProject extends PBXContainer {
 
 	private PBXProject(String name, ImmutableList<PBXTarget> targets, XCConfigurationList buildConfigurationList, List<PBXReference> mainGroupChildren) {
 		this.name = name;
-		this.mainGroup = PBXGroup.builder().name("mainGroup").sourceTree(PBXReference.SourceTree.GROUP).children(mainGroupChildren).build();
+		this.mainGroup = PBXGroup.builder().name("mainGroup").sourceTree(PBXSourceTree.GROUP).children(mainGroupChildren).build();
 		this.targets = Lists.newArrayList(targets);
 		this.buildConfigurationList = buildConfigurationList;
 		this.compatibilityVersion = "Xcode 3.2";
