+++ subprojects/xcode-ide-kit/src/main/java/dev/nokee/xcode/objects/files/PBXReference.java
@@ -15,8 +15,6 @@
  */
 package dev.nokee.xcode.objects.files;
 
-import com.google.common.base.CharMatcher;
-import com.google.common.base.Optional;
 import com.google.common.base.Preconditions;
 import dev.nokee.xcode.objects.PBXContainerItem;
 
@@ -33,9 +31,9 @@ public abstract class PBXReference extends PBXContainerItem {
      * The "base" path of the reference. The absolute path is resolved by prepending the resolved
      * base path.
      */
-    private final SourceTree sourceTree;
+    private final PBXSourceTree sourceTree;
 
-    public PBXReference(String name, @Nullable String path, SourceTree sourceTree) {
+    public PBXReference(String name, @Nullable String path, PBXSourceTree sourceTree) {
         this.name = Preconditions.checkNotNull(name);
         this.path = path;
         this.sourceTree = Preconditions.checkNotNull(sourceTree);
@@ -50,7 +48,7 @@ public abstract class PBXReference extends PBXContainerItem {
         return path;
     }
 
-    public SourceTree getSourceTree() {
+    public PBXSourceTree getSourceTree() {
         return sourceTree;
     }
 
@@ -69,67 +67,4 @@ public abstract class PBXReference extends PBXContainerItem {
             getSourceTree());
     }
 
-    public enum SourceTree {
-        /**
-         * Relative to the path of the group containing this.
-         */
-        GROUP("<group>"),
-
-        /**
-         * Absolute system path.
-         */
-        ABSOLUTE("<absolute>"),
-        /**
-         * Relative to the build setting {@code BUILT_PRODUCTS_DIR}.
-         */
-        BUILT_PRODUCTS_DIR("BUILT_PRODUCTS_DIR"),
-
-        /**
-         * Relative to the build setting {@code SDKROOT}.
-         */
-        SDKROOT("SDKROOT"),
-
-        /**
-         * Relative to the directory containing the project file {@code SOURCE_ROOT}.
-         */
-        SOURCE_ROOT("SOURCE_ROOT"),
-
-        /**
-         * Relative to the Developer content directory inside the Xcode application
-         * (e.g. {@code /Applications/Xcode.app/Contents/Developer}).
-         */
-        DEVELOPER_DIR("DEVELOPER_DIR"),;
-
-        private final String rep;
-
-        SourceTree(String str) {
-            rep = str;
-        }
-
-        /**
-         * Return a sourceTree given a build setting that is typically used as a source tree prefix.
-         *
-         * The build setting may be optionally prefixed by '$' which will be stripped.
-         */
-        public static Optional<SourceTree> fromBuildSetting(String buildSetting) {
-            String data = CharMatcher.is('$').trimLeadingFrom(buildSetting);
-            switch (data) {
-                case "BUILT_PRODUCTS_DIR":
-                    return Optional.of(BUILT_PRODUCTS_DIR);
-                case "SDKROOT":
-                    return Optional.of(SDKROOT);
-                case "SOURCE_ROOT":
-                    return Optional.of(SOURCE_ROOT);
-                case "DEVELOPER_DIR":
-                    return Optional.of(DEVELOPER_DIR);
-                default:
-                    return Optional.absent();
-            }
-        }
-
-        @Override
-        public String toString() {
-            return rep;
-        }
-    }
 }
